// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Plan = {
  FREE: 'FREE',
  STARTER: 'STARTER',
  PRO: 'PRO'
} as const

export type Plan = (typeof Plan)[keyof typeof Plan]


export const Tone = {
  FRIENDLY: 'FRIENDLY',
  BUSINESS: 'BUSINESS',
  LEGAL: 'LEGAL'
} as const

export type Tone = (typeof Tone)[keyof typeof Tone]


export const LetterType = {
  CLAIM: 'CLAIM',
  RESPONSE: 'RESPONSE',
  REMINDER: 'REMINDER'
} as const

export type LetterType = (typeof LetterType)[keyof typeof LetterType]


export const LetterStatus = {
  DRAFT: 'DRAFT',
  SENT: 'SENT',
  OPENED: 'OPENED'
} as const

export type LetterStatus = (typeof LetterStatus)[keyof typeof LetterStatus]